% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refs.R
\name{getLENSData}
\alias{getLENSData}
\title{Function to query Lens.org}
\usage{
getLENSData(token, query)
}
\arguments{
\item{token}{An access key for the lens.org API. Tokens can be obtained by
applying for scholarly API access and creating a token once approved. See
'https://www.lens.org/lens/user/subscriptions#scholar' for further details.}

\item{query}{A search string formulated according to the Lens.org API
documentation: 'https://docs.api.lens.org/request-scholar.html'.}
}
\value{
A summary response. The results are viewable using
'content(data, "text")'. Other details regarding the request (e.g. repsonse
times) can be accessed through the main output.
}
\description{
Function written by lens.org for use of their API.
}
